Moq Properties  and Moq Hierarchy
In a mocked interface, default values ​​come in loose mode, if we want to change the values, we can do this with setup, in strict mode, we must definitely make the setup of all methods. However, if it is a property, default values ​​do not come automatically. We should run the following code to avoid getting a null exception error.
var mockValidator = new Mock<IIdentityValidator>();
   mockValidator.DefaultValue = DefaultValue.Mock;